requirements.txt
fastapi
uvicorn
streamlit
python-dotenv
chromadb
sentence-transformers
torch
google-genai
pydantic
requests